Got player info to render. Working on the dark mode.
